Suppresion des fonctions inutiles Optimisation du code (i++ to ++i) Changement de la classe Random Timer

+++ src/random.h
@@ -4,19 +4,20 @@
 #include <stdlib.h>
 #include <time.h>
 
-class Random{
+class Random
+{
 public:
-    // initialise le générateur de variables aléatoire
-    static void randomize()
-    {
-      srand ( time(NULL) );
-    };
+  // initialise le gï¿½nï¿½rateur de variables alï¿½atoire
+  static void randomize()
+  {
+    srand(time(NULL));
+  };
 
-    // retourne une variable aléatoire entière comprise entre 0 et (borne-1)
-    static long aleatoire(long borne)
-    {
-      return(rand()%borne);
-    };
+  // retourne une variable alï¿½atoire entiï¿½re comprise entre 0 et (borne-1)
+  static int aleatoire(int borne)
+  {
+    return (rand() % borne);
+  };
 };
 
-# endif
+#endif
